requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.piptools
#
black==24.1.1
    # via -r requirements.piptools
cachetools==5.3.2
    # via google-auth
certifi==2024.2.2
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   typer
colorama==0.4.6
    # via typer
distlib==0.3.8
    # via virtualenv
filelock==3.13.1
    # via virtualenv
google-api-core[grpc]==2.16.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.27.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-aiplatform==1.40.0
    # via -r requirements.piptools
google-cloud-bigquery==3.17.1
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.12.0
    # via google-cloud-aiplatform
google-cloud-storage==2.14.0
    # via
    #   -r requirements.piptools
    #   google-cloud-aiplatform
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.62.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-resource-manager
grpcio==1.60.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.60.1
    # via google-api-core
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.3
    # via
    #   pandas
    #   shapely
packaging==23.2
    # via
    #   black
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
pandas==2.2.0
    # via -r requirements.piptools
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via
    #   black
    #   virtualenv
pre-commit==3.6.0
    # via -r requirements.piptools
proto-plus==1.23.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==4.25.2
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pygments==2.17.2
    # via rich
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rich==13.7.0
    # via typer
rsa==4.9
    # via google-auth
shapely==2.0.2
    # via google-cloud-aiplatform
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
typer[all]==0.9.0
    # via -r requirements.piptools
typing-extensions==4.9.0
    # via typer
tzdata==2023.4
    # via pandas
urllib3==2.2.0
    # via requests
virtualenv==20.25.0
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
